Clean up MEF and assembly loader and move to rzls
